fix(imu): emergency fix for the imu msg disorder problem

+++ src/common/time_def.h
@@ -9,8 +9,8 @@ struct UniversalTimeScaleClock {
   using duration = std::chrono::duration<rep, period>;
   using time_point = std::chrono::time_point<UniversalTimeScaleClock>;
   static constexpr bool is_steady = true;
-  static constexpr int f1 = period::den;
-  static constexpr int f2 = std::nano::den / f1;
+  static constexpr int64_t f1 = period::den;
+  static constexpr int64_t f2 = std::nano::den / f1;
 };
 
 using Duration = UniversalTimeScaleClock::duration;
